refactor: extract components (Header, Hero, Skills, Projects, Experience, Contact, Footer) and simplify App\n\n💘 Generated with Crush\nCo-Authored-By: Crush <crush@charm.land>
