+++ src/components/Footer.tsx
@@ -1,0 +1,8 @@
+export function Footer() {
+  return (
+    <footer className="container mb-10 text-center text-xs font-bold">
+      <span style={{ color: '#415E72' }}>© 2025 Javier Sanchez Daza.</span>{' '}
+      <span style={{ color: '#6f4b84' }}>Built with React + Tailwind.</span>
+    </footer>
+  )
+}
